change types for cloudness

+++ src/main/java/ru/barabanra/yandex/dto/item/weather/Fact.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 public class Fact {
 
-    private Long cloudness;
+    private Double cloudness;
 
     private String condition;
